third_party/blink/web_tests/fast/lists/list-style-position-inside-expected.txt

PASS All items in floating ul has the expected height.
PASS Width of first item in floating ul is affected by list marker as expected.
PASS All items in floating ol has the expected height.
PASS Width of first item in floating ol is affected by list marker as expected.
PASS All items in block ul has the expected height.
PASS Width of first item in block ul is affected by list marker as expected.
PASS All items in block ol has the expected height.
PASS Width of first item in block ol is affected by list marker as expected.
PASS successfullyParsed is true

TEST COMPLETE
